requirements.txt
jinja2
pyyaml
requests
jinja2-ansible-filters
